Added a system to serialize DDRings and DDFunctions
Added a new miscellaneous class (SerializableObject) that provides all
the functionality for serializing (writing on a file) objects. It is
based on the package pickle and save the initialization of an object to
reconstruct it when loaded.

To use this functionality, a class must only inherit from
SerializableObject and initialize the class in the __init__ method
providing the arguments received in the __init__method.

Added the functionaliti of SerializableObject to the classes of the file
ddFunction. Now, DDRings, ParametrizedDDRings and DDFunctions can be
serialized into files.

In particular, the serialization for DDFunctions has an extra parameter
that allows the user to save all the initial conditions that have been
already computed.