Changed method for getting the systems
Previously, the method get_matrix_add and get_matrix_mul
allowed to get the ansatz system for the usual computations.

Since the code was fairly similar to the building of the ansatz systems
for simple computations, we changed the meaning of these functions
allowing now to build systems of any size, inhomogenous or not. This
required a small adaptation of the methods
that computed the annihilating operator (the two steps method such as
_compute_add_solution) to include the
size of the system and the homogenous requirement.

These new methods need to be adapted in all inherited classes (like
fullLazyOperator).