Added the control of leading coefficients
Now the set S for the leading coefficients can be given by the user for simple computations. This input, currently, has to be a list of elements to generate a multiplicatively closed set. Then we check if the minimal set is included in this new set and then we perform the usual computations.